Implementacion login, registro, mascotas pendientes de adopcion y total mascotas adoptadas

+++ TuMascota/TuMascota.AccesoDatos/Modulos/Core/Usuarios/IRepoUsuarios.cs
@@ -4,16 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TuMascota.AccesoDatos.DataModels;
-using UIBASOFT.Base.Comunes.ResultOperation;
 
 namespace TuMascota.AccesoDatos.Modulos.Core.Usuarios
 {
     public interface IRepoUsuarios : IDisposable
     {
-        Task<Usuario> Obtener(int id);
+        Task<Usuario> Obtener(long id);
         Task<int> Crear(Usuario entity);
         Task<Usuario> Editar(Usuario dto);
-        Task<int> Eliminar(int id);
+        Task<int> Eliminar(long id);
         IUnitOfWork UnitOfWork { get; }
     }
 }
